+++ packages/html/src/define/ui/time-slider.ts
@@ -1,6 +1,8 @@
 import { SliderBufferElement } from '../../ui/slider/slider-buffer-element';
 import { SliderFillElement } from '../../ui/slider/slider-fill-element';
+import { SliderPreviewElement } from '../../ui/slider/slider-preview-element';
 import { SliderThumbElement } from '../../ui/slider/slider-thumb-element';
+import { SliderThumbnailElement } from '../../ui/slider/slider-thumbnail-element';
 import { SliderTrackElement } from '../../ui/slider/slider-track-element';
 import { SliderValueElement } from '../../ui/slider/slider-value-element';
 import { TimeSliderElement } from '../../ui/time-slider/time-slider-element';
@@ -10,7 +12,9 @@ import { safeDefine } from '../safe-define';
 safeDefine(TimeSliderElement);
 safeDefine(SliderBufferElement);
 safeDefine(SliderFillElement);
+safeDefine(SliderPreviewElement);
 safeDefine(SliderThumbElement);
+safeDefine(SliderThumbnailElement);
 safeDefine(SliderTrackElement);
 safeDefine(SliderValueElement);