+++ packages/react/src/ui/slider/index.parts.ts
@@ -1,6 +1,8 @@
 export { SliderBuffer as Buffer, type SliderBufferProps as BufferProps } from './slider-buffer';
 export { SliderFill as Fill, type SliderFillProps as FillProps } from './slider-fill';
+export { SliderPreview as Preview, type SliderPreviewProps as PreviewProps } from './slider-preview';
 export { SliderRoot as Root, type SliderRootProps as RootProps } from './slider-root';
 export { SliderThumb as Thumb, type SliderThumbProps as ThumbProps } from './slider-thumb';
+export { SliderThumbnail as Thumbnail, type SliderThumbnailProps as ThumbnailProps } from './slider-thumbnail';
 export { SliderTrack as Track, type SliderTrackProps as TrackProps } from './slider-track';
 export { SliderValue as Value, type SliderValueProps as ValueProps } from './slider-value';